Remove duplicate helper function and update references.
