+++ examples/__init__.py
@@ -2,4 +2,3 @@
 #  SPDX-License-Identifier: Apache-2.0
 
 from .example_config import ExampleConfig
-from .helper_functions import get_account_from_config
