+++ examples/example_config.py
@@ -71,20 +71,6 @@ class ExampleConfig:
         return ExampleConfig._net_to_env_name.get(net)
 
     @staticmethod
-    def get_accounts_config(local_node=True):
-        if local_node:
-            a, p = "0x00bd138abd70e2f00903268f3db08f2d25677c9e", "node0"
-            a1, p1 = "0x068ed00cf0441e4829d9784fcbe7b9e26d4bd8d0", "secret"
-        else:
-            a, p = get_variable_value('PARITY_ADDRESS'), get_variable_value('PARITY_PASSWORD')
-            a1, p1 = get_variable_value('PARITY_ADDRESS1'), get_variable_value('PARITY_PASSWORD1')
-
-        return {
-           "parity.address": a, "parity.password": p,
-           "parity.address1": a1, "parity.password1": p1
-        }
-
-    @staticmethod
     def get_base_config():
         config = {
             "keeper-contracts": {
@@ -101,8 +87,6 @@ class ExampleConfig:
                 "downloads.path": "consume-downloads"
             }
         }
-        # Accounts info is in env vars, need to deprecate the config accounts options
-        # config['keeper-contracts'].update(ExampleConfig.get_accounts_config(local_node))
         return config
 
     @staticmethod
