--- examples/helper_functions.py
@@ -1,22 +1,0 @@
-#  Copyright 2018 Ocean Protocol Foundation
-#  SPDX-License-Identifier: Apache-2.0
-
-from squid_py.accounts.account import Account
-from squid_py.keeper import Keeper
-from squid_py.keeper.web3_provider import Web3Provider
-
-
-def get_account_from_config(config, config_account_key, config_account_password_key):
-    address = None
-    if config.has_option('keeper-contracts', config_account_key):
-        address = config.get('keeper-contracts', config_account_key)
-        address = Web3Provider.get_web3().toChecksumAddress(address) if address else None
-
-    if not (address and address in Keeper.get_instance().accounts):
-        return None
-
-    password = None
-    if address and config.has_option('keeper-contracts', config_account_password_key):
-        password = config.get('keeper-contracts', config_account_password_key)
-
-    return Account(address, password)
